requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.4.0
Werkzeug==3.0.3
gunicorn==22.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.4.0
fpdf2==2.7.9
requests==2.32.3
